Extract sync delta registry
