+++ tests/test-sync.js
@@ -42,6 +42,7 @@ await import('../js/settings.js');
   const syncDisableCleanupSrc = await fetchWithRetry('js/sync-disable-cleanup.js');
   const syncSchemaSrc = await fetchWithRetry('js/sync-schema.js');
   const syncDeltaSrc = await fetchWithRetry('js/sync-delta.js');
+  const syncDeltaRegistrySrc = await fetchWithRetry('js/sync-delta-registry.js');
   const syncTombstonesSrc = await fetchWithRetry('js/sync-tombstones.js');
   const syncMessengerSrc = await fetchWithRetry('js/sync-messenger.js');
   const syncEnvironmentSrc = await fetchWithRetry('js/sync-environment.js');
@@ -67,7 +68,7 @@ await import('../js/settings.js');
   const stylesSrc = await fetchWithRetry('styles.css');
   const themeExtraSrc = await fetchWithRetry('themes-extra.css');
   const serviceWorkerSrc = await fetchWithRetry('service-worker.js');
-  const deltaSearchSrc = `${syncSrc}\n${syncPushSrc}\n${syncReconcileSrc}\n${syncPullSrc}\n${syncCutoverSrc}\n${syncDeltaSrc}\n${syncDiagnosticsSrc}\n${syncDiagnoseUiSrc}\n${syncWindowBindingsSrc}`;
+  const deltaSearchSrc = `${syncSrc}\n${syncPushSrc}\n${syncReconcileSrc}\n${syncPullSrc}\n${syncCutoverSrc}\n${syncDeltaSrc}\n${syncDeltaRegistrySrc}\n${syncDiagnosticsSrc}\n${syncDiagnoseUiSrc}\n${syncWindowBindingsSrc}`;
   const exportBlockIncludes = (src, names) => [...src.matchAll(/export\s+\{([^}]*)\};/g)]
     .some(([, block]) => names.every(name => new RegExp(`\\b${name}\\b`).test(block)));
 
@@ -173,12 +174,21 @@ await import('../js/settings.js');
     serviceWorkerSrc.includes("'/js/sync-apply.js'"));
   assert('sync-delta.js owns per-row delta helpers',
     syncSrc.includes("from './sync-delta.js'")
-      && syncDeltaSrc.includes('export const DELTA_ARRAYS')
+      && syncDeltaSrc.includes("from './sync-delta-registry.js'")
       && syncDeltaSrc.includes('export async function _planArrayDelta')
       && syncDeltaSrc.includes('export async function _mergeItemRowsIntoImported')
       && syncDeltaSrc.includes('export function getDeltaCutoverReadiness'));
   assert('service worker precaches sync-delta.js',
     serviceWorkerSrc.includes("'/js/sync-delta.js'"));
+  assert('sync-delta-registry.js owns delta surfaces and identity config',
+    syncDeltaRegistrySrc.includes('export const DELTA_ARRAYS')
+      && syncDeltaRegistrySrc.includes('export const DELTA_MAPS')
+      && syncDeltaRegistrySrc.includes('export const DELTA_SCALARS')
+      && syncDeltaRegistrySrc.includes('export const DELTA_ARRAY_CONFIG')
+      && syncDeltaRegistrySrc.includes('export const DELTA_MAP_CONFIG')
+      && syncDeltaRegistrySrc.includes('export function _isAllowlistSafeId'));
+  assert('service worker precaches sync-delta-registry.js',
+    serviceWorkerSrc.includes("'/js/sync-delta-registry.js'"));
   assert('sync-tombstones.js owns remote profile delete helpers',
     syncSrc.includes("from './sync-tombstones.js'")
       && syncTombstonesSrc.includes('export async function deleteProfileFromRelay')
@@ -555,15 +565,15 @@ await import('../js/settings.js');
 
   // DELTA_ARRAYS list (high-velocity arrays)
   assert('DELTA_ARRAYS includes sunSessions + lightDevices',
-    /DELTA_ARRAYS\s*=\s*\[[\s\S]{0,400}'sunSessions'[\s\S]{0,400}'lightDevices'/.test(syncDeltaSrc));
+    /DELTA_ARRAYS\s*=\s*\[[\s\S]{0,400}'sunSessions'[\s\S]{0,400}'lightDevices'/.test(syncDeltaRegistrySrc));
   assert('DELTA_ARRAYS includes entries + notes (high-importance lab data)',
-    /DELTA_ARRAYS\s*=\s*\[[\s\S]{0,800}'entries'[\s\S]{0,400}'notes'/.test(syncDeltaSrc));
+    /DELTA_ARRAYS\s*=\s*\[[\s\S]{0,800}'entries'[\s\S]{0,400}'notes'/.test(syncDeltaRegistrySrc));
 
   // Push-side plan/apply contract
   assert('_planArrayDelta diffs against last-pushed snapshot',
     /_planArrayDelta[\s\S]{0,1200}_readDeltaSnapshot\(profileId,\s*arrayName\)[\s\S]{0,1200}prev\[itemId\]\s*===\s*hash/.test(syncDeltaSrc));
   assert('_planArrayDelta validates itemId allowlist (defence-in-depth)',
-    /\^\[a-zA-Z0-9_\.-\]\+\$/.test(syncDeltaSrc));
+    /\^\[a-zA-Z0-9_\.-\]\+\$/.test(syncDeltaRegistrySrc));
   assert('_planArrayDelta gzip-compresses payloads >256 bytes',
     /json\.length > 256[\s\S]{0,200}GZ\|v1\|/.test(syncDeltaSrc));
   assert('_planArrayDelta emits tombstones for items removed since last push',
@@ -2002,7 +2012,7 @@ await import('../js/settings.js');
   // in a sibling const + report a false positive.
   const inList = (constName, entry) => {
     const re = new RegExp(`const ${constName}\\s*=\\s*\\[([\\s\\S]*?)\\]`, 'm');
-    const m = syncDeltaSrc.match(re);
+    const m = syncDeltaRegistrySrc.match(re);
     if (!m) return false;
     return m[1].includes(`'${entry}'`);
   };
