+++ tests/test-v1-6-shipped.js
@@ -197,8 +197,8 @@ const _origProfileSex = window._labState ? window._labState.profileSex : null;
     // flag). Without this, under v4 payloads the supersession-
     // generated tombstones (which ride _deleted in v3) never reach
     // peers, and paired devices retain stale measurements forever.
-    const syncDeltaSrc = fetchSrc('js/sync-delta.js');
-    const cfgBlock = syncDeltaSrc.split('DELTA_ARRAY_CONFIG')[1] || '';
+    const syncDeltaRegistrySrc = fetchSrc('js/sync-delta-registry.js');
+    const cfgBlock = syncDeltaRegistrySrc.split('DELTA_ARRAY_CONFIG')[1] || '';
     const lmCfgMatch = cfgBlock.match(/lightMeasurements:\s*\{[\s\S]{0,300}?\}/);
     assert('sync-delta.js: lightMeasurements has NO noTombstones (Phase 2 propagation)',
       !lmCfgMatch || !/noTombstones:\s*true/.test(lmCfgMatch[0]),
